% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_case_study_data.R
\name{extract_case_study_data}
\alias{extract_case_study_data}
\title{Extract the case study dataset}
\usage{
extract_case_study_data(exdir = ".", files = NULL)
}
\arguments{
\item{exdir}{A string giving the path to the extraction directory. Passed to
\code{\link[utils]{unzip}}.}

\item{files}{A character vector of files to extract, or \code{NULL} to 
extract all files.  Passed to \code{\link[utils]{unzip}}.}
}
\value{
A character vector of the extracted file paths is invisibly returned.
}
\description{
Extracts the data files used by the case study from the zip archive.
}
\examples{
extract_case_study_data(tempfile("flippant"))
}
\seealso{
\code{\link[utils]{unzip}}
}
\author{
Richard Cotton
}
