% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flippant-package.R
\docType{package}
\name{flippant}
\alias{flippant}
\title{flippant}
\description{
A package for the analysis of data provided by dithionite 
scramblase assays.
}
\author{
Johannes Graumann
}
\keyword{package}
