% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_felix_gx_output.R
\name{parse_felix_gx_output}
\alias{parse_felix_gx_output}
\title{parse_felix_gx_output}
\usage{
parse_felix_gx_output(x = NULL)
}
\arguments{
\item{x}{\code{\link{character}} vector resulting from 
\code{\link{readLines}} of the corresponding data file.}
}
\value{
See \code{\link{parse_fluorimeter_output}}.
}
\description{
Parse spectra from files provided by a QuantaMaster fluorimeter 
(Photon Technology International, Inc., Edison, New Jersey) using 
\code{FelixGX v4.1}
}
\details{
A helper function to \code{\link{parse_fluorimeter_output}}.
}
\seealso{
\code{\link{parse_fluorimeter_output}},
\code{\link[flippant]{parse_easyspec_output}},
\code{\link[flippant]{parse_felix_32_output}},
\code{\link[flippant]{parse_FluorS_Essence_3.8_output}},
\code{\link[flippant]{parse_manual_output}}
}
\author{
Johannes Graumann
}
\keyword{IO}
\keyword{file}
\keyword{manip}
