% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_FluorS_Essence_3.8_output.R
\name{parse_FluorS_Essence_3.8_output}
\alias{parse_FluorS_Essence_3.8_output}
\title{parse_FluorS_Essence_3.8_output}
\usage{
parse_FluorS_Essence_3.8_output(x = NULL)
}
\arguments{
\item{x}{\code{\link{character}} vector resulting from 
\code{\link{readLines}} of the corresponding data file.}
}
\value{
See \code{\link{parse_fluorimeter_output}}.
}
\description{
Parse spectra from files provided by a Horiba fluorimeter 
(HORIBA Europe GmbH, Oberursel, Germany) using \code{FluorS Essence v3.8}
}
\details{
A helper function to \code{\link{parse_fluorimeter_output}}.
}
\seealso{
\code{\link{parse_fluorimeter_output}},
\code{\link[flippant]{parse_easyspec_output}},
\code{\link{parse_felix_32_output}},
\code{\link[flippant]{parse_felix_gx_output}},
\code{\link[flippant]{parse_manual_output}}
}
\author{
Johannes Graumann
}
\keyword{IO}
\keyword{file}
\keyword{manip}
