% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_manual_output.R
\name{parse_manual_output}
\alias{parse_manual_output}
\title{parse_manual_output}
\usage{
parse_manual_output(x = NULL)
}
\arguments{
\item{x}{\code{\link{character}} vector resulting from 
\code{\link{readLines}} of the corresponding data file.}
}
\value{
See \code{\link{parse_fluorimeter_output}}.
}
\description{
Parse spectra from files provided in a manually assembled 
format.
}
\details{
A helper function to \code{\link{parse_fluorimeter_output}}.
The file in question is supposed to contain a tab-delimited table
with the columns \code{Time (sec)} and \code{Fluorescense Intensity}.
}
\seealso{
\code{\link{parse_fluorimeter_output}},
\code{\link[flippant]{parse_easyspec_output}},
\code{\link[flippant]{parse_felix_gx_output}}, 
\code{\link[flippant]{parse_felix_32_output}}
}
\author{
Johannes Graumann
}
\keyword{IO}
\keyword{file}
\keyword{manip}
